Moving pdf KeyboardInfo sidebar, ui and viewer
