+++ client/app/reader/PdfSidebar.jsx
